% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_c.R
\name{print_matrix}
\alias{print_matrix}
\title{Example function to call C from R}
\usage{
print_matrix(data)
}
\arguments{
\item{data}{A N x M data matrix}
}
\value{
A list of length 3, containing (a) The number of rows in `data`; (b) the 
number of columns in `data`
}
\description{
Takes a matrix and prints it.
}
\details{
The input argument is passed as a pointer to C. To get a return value,
change the data pointed to by this pointer in C.
}
\examples{
N <- 8
data <- matrix(rnorm(N * 2), ncol = 2)
# Print a distance matrix
distance_matrix <- as.matrix(dist(data))
print_matrix(distance_matrix)

}
